# Memoize Block.DecoderSpec