`DecoderSpec` is called many times throughout terraform, and in the process of checking for BlocktoAttr fixups, it may be called n^2 times recursively. While this is not normally a problem for most common resources, with deeply nested schemas this can cause serious performance issues. 

Because of the delicate nature of the BlockToAttr fixup code in dealing with legacy provider schemas, it would be preferable to leave that code as-is for he time being if at all possible. Rather than refactor the BlockToAttr code ((if that's possible at all without subtle breaking changes), we have an easy performance gain here by caching the `Block.DecoderSpec` calls.

The benchmark shown here produced the following results:
```
name               old time/op    new time/op    delta
FixUpBlockAttrs-4     9.13s ± 1%     0.84s ± 3%  -90.78%  (p=0.000 n=10+8)

name               old alloc/op   new alloc/op   delta
FixUpBlockAttrs-4    6.02GB ± 0%    0.41GB ± 0%  -93.24%  (p=0.000 n=10+9)

name               old allocs/op  new allocs/op  delta
FixUpBlockAttrs-4     54.7M ± 0%      3.7M ± 0%  -93.21%  (p=0.000 n=10+9)
```

and testing a simple configuration with a `aws_wafv2_web_acl` resource in the CLI produced roughly a 98% percent speedup.

The less attractive part of this change is the addition of the package-level `specCache`. Normally one would pair the volatile field in the struct with its own mutex, but due to the fact that `Block` is assigned by value to `NestedBlock`, and often copied by assignment, there is no way to insert a mutex without introducing breaking changes in the public API. The package-level mutex used by `specCache` will have no impact, since the number of calls to `DecoderSpec` will be significantly reduced.

Fixes #25889