@jbardin 2020-10-13
While the package level mutex is simpler, This needs to be refactored to use a separate cache. Storing the value within the `Block` is still technically unsafe, because one could copy the `Block` value while the `decoderSpec` field is being written.