@jbardin 2020-10-14
@alisdair, yes the  "mutability" is odd here, but the `Block` type should be used as if it were an immutable value once the full tree is constructed. It's a representation of a schema, which cannot change while it's being used, so any modifications during execution would be an error. It didn't seem worth the effort to generate a recursive hash for the block tree, but that is technically possible if there are still concerns.