@alisdair 2020-10-14
I don't have any concerns! The absolute most I'd suggest is adding your above comment about how `Block` is used to the source near the cache struct for any future readers who worry needlessly about the same thing. But that's very optional.

✅